fix: give the call detail component a much needed cleanup
