+++ clients/web/src/components/calls/detail-pane/traces/popover/content.tsx
