+++ clients/web/src/components/calls/detail-pane/traces/popover/tool-tip/content.tsx
