+++ clients/web/src/components/calls/detail-pane/traces/slice.tsx
