--- clients/web/src/components/calls/detail-pane/view.tsx
@@ -1,87 +1,0 @@
-import { For, JSXElement, Setter } from "solid-js";
-import { Trace } from "./trace";
-import { Args } from "./args";
-import clsx from "clsx";
-import { TreeEntry } from "~/lib/span/get-span-children";
-import { Metadata_Level } from "~/lib/proto/common";
-import { FilterToggle } from "~/components/filter-toggle";
-import { Span } from "~/lib/connection/monitor";
-import { RootLink } from "./root-link";
-
-export function View(props: {
-  name: string;
-  minLevel: Metadata_Level;
-  setMinLevel: Setter<Metadata_Level>;
-  hasError: boolean | null;
-  parentId?: bigint;
-  rootSpan?: Span;
-  spanChildren: TreeEntry[];
-  valuesSectionTitle: string;
-  values: (string | object)[];
-  children: JSXElement;
-}) {
-  const closedSpans = () =>
-    props.spanChildren.filter((s) => s.span.closedAt > 0);
-
-  const durations = () => {
-    return {
-      start: Math.min(...closedSpans().map((s) => s.span.createdAt)),
-      end: Math.max(...closedSpans().map((s) => s.span.closedAt)),
-      shortest: Math.min(...closedSpans().map((s) => s.span.duration)),
-      longest: Math.max(...closedSpans().map((s) => s.span.duration)),
-    };
-  };
-
-  return (
-    <div class="h-full overflow-auto grid gap-4 content-start border-l border-gray-800 min-w-[420px]">
-      <div class="pt-4 px-4">
-        <RootLink rootSpan={props.rootSpan} />
-        <h2 class={clsx("text-2xl", props.hasError ? "text-red-400" : "")}>
-          {props.name}
-        </h2>
-      </div>
-      <div class="grid gap-2">
-        <h2 class="text-xl p-4">{props.valuesSectionTitle}</h2>
-        <table>
-          <tbody>
-            <Args args={props.values} />
-          </tbody>
-        </table>
-      </div>
-      {props.children}
-      <div class="grid gap-2">
-        <h2 class="text-xl p-4">Trace</h2>
-        <div class="items-center flex justify-end h-toolbar p-2 gap-4">
-          <FilterToggle
-            defaultPressed={props.minLevel > Metadata_Level.DEBUG}
-            aria-label="attributes"
-            changeHandler={() =>
-              props.setMinLevel((prev) =>
-                prev == Metadata_Level.TRACE
-                  ? Metadata_Level.DEBUG
-                  : Metadata_Level.TRACE,
-              )
-            }
-          >
-            <span>Show trace level</span>
-          </FilterToggle>
-        </div>
-        <table>
-          <thead>
-            <tr class="uppercase border-gray-800 border-b">
-              <th>Name</th>
-              <th>Waterfall</th>
-            </tr>
-          </thead>
-          <tbody>
-            <For each={props.spanChildren}>
-              {({ depth, span }) => (
-                <Trace depth={depth} span={span} durations={durations()} />
-              )}
-            </For>
-          </tbody>
-        </table>
-      </div>
-    </div>
-  );
-}
